Refactor ecosystem package structure, rename classes, and add PurchaseRequestService and ShipmentStatus
