+++ Procurement_Ecosystem/src/enums/ShipmentStatus.java
@@ -1,0 +1,8 @@
+package enums;
+
+/**
+ * @author tisaac
+ */
+public enum ShipmentStatus {
+    PLACED, IN_TRANSIT, DELIVERED
+}
