+++ Procurement_Ecosystem/src/model/ecosystem/Ecosystem.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Ecosystem;
+package model.ecosystem;
 
 //import Business.Person.PersonDirectory;
